20180221_cleaning
removed camera lucida assisted graphic as it is an outdated method to
make a graphic and would not pertain to current institute work flows

started new folder entitled no longer technologically relevant to dump
old terms like the detectors not based in CCD or CMOS technology…

started new folder for technology we do not have (or are likely to in
the next five years) such as a neutron microscope.

added montage of images to reflect stills of a timelapse etc.

sample prep changed some structure. created fixed sample parent class
to parallel construction match unfixed category.